+++ src/features/TextToImage/models/SDXL/client/exports.ts
@@ -8,7 +8,7 @@ import ShimmedStabilityAIClient from '@/library/ShimmedStabilityAIClient/definit
 
 import type {
   Output,
-} from '@/features/TextToImage/types';
+} from '@/features/TextToImage/types/exports';
 
 import {
   Server,
